0.3 Easing
Added:
30(plus linear) easing methods for tweening. By default tween will use linear, to use diffrent ease enum of choice must be passed to choosen tween method.

ColorTween.

CanvasGroupTween(alpha).

Class for changing transforms individual axis(name will be changed to better suit its needs and more functionality will be added).

Changed:
Delays are not longer stored in dictionary, so in order to cancel them, I set delay as variable and pass it to the  cancel method.